fix type of event

+++ src/Os/Scheduler/Os_Types.h
@@ -31,7 +31,7 @@ typedef struct tsOsTaskCfg
 
 typedef struct tsOsTaskTCB
 {
-  uint8_t        ulEvent;
+  uint32_t       ulEvent;
   uint32_t       ulAlarm;
 }tsOsTaskTCB;
